Initial port of Phase 2 source code into our ecosystem.
After spending quite a bit of time at both semester's specs and porting
source code, I've come to the conclusion that we need to address
terminology and rename some things as we pull in code from the Spring code
base.

Joel's EnumRegion, for our intents and purposes, groups political units
(countries and US states) into geographical regions for book keeping and
display purposes.  This will be the master definition of the term
'Region'.  In other words, the term 'Region' will refer to a region of the
world per the new spec, aggregating multiple agricultural units.  Each
region may be composed of multiple countries or states.

The Spring project divided the Earth's surface into units called LandTiles
for the purposes of planting and harvesting crops.  Their architecture
organizes tiles into Countries, but they were't required to break the US
down into individual states or aggregate countries into Regions.
Correspondingly, their basic book keeping unit (and class name) was by
Country.  They used sets of lat/lon coordinates called MapPoints to
describe polygons corresponding to cartographic regions of the Earth's
surface.  They loosely referred to each collection of points as a
'Region', using an interface by that name.  The borders of each Country
were described by one of these cartographic regions.  Each LandTile is
assigned to a country using a simple iterative test for intersection of
the geographic center of the tile with each country's geographic area.

I have refactored the Spring code from the geographic Region to
'GeographicArea' to differentiate it from the Fall semester's concept of a
Region.  A new class hierarchy represents Regions, Countries, States.

Because we now need to keep track of the production of each of the 50 US
states and aggregate states and countries into regions, I have refactored
the Country class into a hierarchy based on the idea of an
AgriculturalUnit:

   AbstractAgriculturalUnit - The basic data structure for keeping track
of data (formerly AbstractCountry).
   AgriculturalUnit - The basic unit for which we track / update
agricultural data (i.e a country or state, was Country).
   State (extends AgriculturalUnit) - A convenience class for readability.
   Country (extends AgriculturalUnit) - A convenience class for
readability.
   Region - a collection of AgriculturalUnits for each of the geographic
regions used in the game (See EnumRegion)
   World - a collection of AgriculturalUnits

I haven't done anything with the readers.  The country reader still reads
countries, only it now generically calls them an Agricultural unit.  It
does not collect countries or states into Regions yet.  We may choose to
describe the geography of Regions similarly to GeographicAreas using these
polygons, and test for intersection between Country/State and Region and
polygons (GeographicAreas) in order to determine which country goes into
which Region bucket for book-keeping purposes.  The alternative is to
construct a list of countries belonging to each region.

At this point I have all of last semester's code ported into our hierarchy
using our EnumFood and other new classes.  It all compiles clean,
including the various refactoring that I have done.  I haven't added the
math for the new crop/food types, but that's a next step along with
mapping Regions.  I have left it in its original package hierarchy rather
than dropping it all into starvationevasion, so that we can decide if it
needs to be reorganized.