Added adaptive method of picking cards to draft.
When a new draft object is created, the information about what happened during the last turn is evaluated and if things improved, the same type of policy, food or region is likely to be picked again. If things got worse, it is not as likely that those same items will be picked again.