Added in EnumGameState, represents various game states
Meant to keep track of when cards can be played. Policie Cards can only be played during the planning phase, while something like the Filibuster Action-card (which forces a policy that is being voted on back to the hand of the proposer, without it passing) can only be played during the voting phase.