Merges John Clark's simulator updates up for commit d31db7b. Request ID: 27
John Clark's comment:
All issues with the main branch should be resolved.
Added to Model.java: placeCrops() which iterates through each region, placing crops on landtiles.

Added to Region.java: getCropArea() returns area used by the requested crop.
getNumTiles() returns the number of landTiles in the region.

If you would like to test, to see that crops are actually being placed, there is code at the bottom of placeCrops that is commented out currently. Also, there is no need to start up the entire game in order to test this, as there is a main method within Model.java.