disabled ConvertIntelligence card
That card needs a bunch of work to be functional, such as being able to, mid-drafting, call the server and ask for a player's deck, and to be able to examine it. AFAIK, the current setup only has information coming in from the server at the the start/ends of each phase, and the structure is not designed to be able to poll information from the sever in between those phases.